// src/app/Models/Experts Model/Medco Experts.ts
export class MedcoExperts{

  ID:string;
  NamePrefix: string;
  LastName: string;
  FirstName:string;
  MiddleName: string;
  NameSuffix: string;
  FullName:string;
  Gender:string;
  AddressLine1: string;
  AddressLine2: string;
  County: string;
  Town: string;
  PostCode: string;
  PhoneNo: string;
  AlternatePhoneNo: string;
  FaxNo:string;
  Email:string;
  MobileNumber:string;
  GeneralEmail: string;
  QualificationDescription:any;
  FurtherSpeciality:any;
  SpecialityID:string;
  SignaturePath:any;
  CVPath:string;
  AccreditationCertificatePath: string;
  ContractDocumentPath: string;
  MedcoCertificatePath:string;
  InsuranceDocumentPath: string;
  IcoPath:string;
  PINCode:string;
  GmcHcpc: string;
  Indeminity:string;
  Ico:string;
  MedcoNo: string;
  AccreditationNo: string;
  ExpertContractType: string;
  AmountRate: string;
  BankName: string;
  AccountTitle: string;
  AccountNo: string;
  SortCode:string;
  ExpertTypeID:string;
  DueDays: string;
  ExpertSpecialityID: string;
  IsVATRegistered:any;
  IsLocumAgencyExpert:any;
  IsAvailableForReferrer: any;
  IsAgreeWithTermAndCondition: any;
  IsAvailableForGoPrivate: any;
  IsAvailableForIdox: any;
  OnPanel: string;
  IACharges:string;
  DCCharges:string;
  PerSessionCharges:string;
  DNACharges: string;
  Status:string;
  CreatedBy:string;
  CreatedDate:string;
  signaturefile: any;
  cVfile: any;
  accCertificatefile: any;
  contractfile:any;
  icOfile:any;
  insurancefile:any;
  medcoCertificatefile: any;
  

}

export class QualificationData{
  QualificationDescription:any;
}

export class FurtherSpecialityData{
  FurtherSpeciality:any;
}